pkg/narinfo: move from pkg/nar/narinfo
There's not much of a reason to put this inside the nar package.
These files can be useful even without a .nar file around.
