+++ pkg/narinfo/check.go
