+++ pkg/narinfo/fingerprint.go
