+++ pkg/narinfo/signature/secret_key.go
