+++ pkg/narinfo/signature/util.go
